% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_political_juv.R
\docType{data}
\name{data_political_juv}
\alias{data_political_juv}
\title{Data for anallysis of political interests of the youth}
\format{Lista con la descripción de las :
\describe{
\item{P38}{Cursando o no cursando el colegio }
\item{P52_x}{Confianza en las intittuciones, enumeración de 10 instituciones desde P52_1 hasta P52_10 }
}}
\usage{
data_political_juv
}
\description{
A data \code{tibble} that is for multiples analysis
}
\keyword{datasets}
