% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{data_read}
\alias{data_read}
\title{Read CSV duplicate}
\usage{
data_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{tibble}
}
\description{
Read CSV duplicate
}
\examples{
csv= system.file("extdata","sample_data.csv", package = "compol")
sample_read(csv)
}
