% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audiencias_pag.R
\name{get_audiencias_pag}
\alias{get_audiencias_pag}
\title{Obtención de audiencias}
\usage{
get_audiencias_pag(ini_page, num_page)
}
\arguments{
\item{ini_page}{Es el número de pagina inicial de audiencias}

\item{num_page}{Es el número de pagina final de audiencias}
}
\value{
Un \dataframe con las audiencias estructuradas
}
\description{
Obtención de audiencias
}
\examples{
get_audiencias_pag(num_page=3)

}
