% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello CpyE function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi to.}
}
\value{
the output from \code{\link{print}}
}
\description{
My Hello CpyE function
}
\examples{
hello("Sergio" )
\dontrun{
hello("Steve")
}

}
