% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobACT.R
\name{lobACT}
\alias{lobACT}
\title{Actualización resumen de datos del lobby}
\usage{
lobACT()
}
\arguments{
\item{resumen_lobby}{Muestra un resumen de el numero e instituciones, cargos_pasivos, cargos_activos y audiencias.}
}
\value{
un resumen de datos de instituciones, cargos_pasivos, cargos_activos y audiencias.
}
\description{
Actualización resumen de datos del lobby
}
\examples{
lobACT(resumen_lobby=1)
}
