% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political_jus_read.R
\name{political_read_data}
\alias{political_read_data}
\title{Read CSV duplicate (political data juv)}
\usage{
political_read_data(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{tibble}
}
\description{
Read CSV duplicate (political data juv)
}
\examples{
data_political= system.file("extdata","data_political_juv.csv",package = "compol")
political_read_data(csv)
}
