% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Title}
\usage{
printer(r, x)
}
\arguments{
\item{r}{what you want in the second column}

\item{x}{what you want in the first column}
}
\value{
A tibble
}
\description{
Title
}
\examples{
printer(x=rnorm(10), r= rnorm(10))
}
