% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Data for basic analysis}
\format{A list with 7 elements, which are:
\describe{
\item{x}{random normal data }
\item{y}{outcome based on \code{x} }
}}
\usage{
sample_data
}
\description{
A simple \code{tibble} that is for simple analysis
}
\keyword{datasets}
