% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audiencias_pag.R
\name{get_audiencias_pag}
\alias{get_audiencias_pag}
\title{Obtención de audiencias}
\usage{
get_audiencias_pag(num_page)
}
\arguments{
\item{num_page}{Es el número de paginas de audiencias a retornar, la API devuelve de a 10 audiencias por pagina}
}
\value{
Un \dataframe con las audiencias estructuradas
}
\description{
Obtención de audiencias
}
\examples{
get_audiencias_pag(num_page=3)

}
