% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobbyR-package.R
\docType{package}
\name{lobbyR-package}
\alias{lobbyR}
\alias{lobbyR-package}
\title{lobbyR: Functions to Get Federal Lobbying Disclosures}
\description{
lobbyR allows users to query the Senate federal lobbying disclosures database \url{https://lda.senate.gov/api/redoc/v1/} to find out how much companies and other entities are spending to lobby Congress and the federal government. It allows for search terms such as keywords, time periods and entity names. It then attempts to clean, or at least flag, filings that could provide incorrect results when seeking to answer the question: How much is being spent on lobbying our Congress and the administration and what issues do they care about?
}
\author{
\strong{Maintainer}: Chris Cioffi \email{chrisjcioffi@gmail.com}

Authors:
\itemize{
  \item Aarushi Sahejpal \email{aarushisahejpal@gmail.com} [contributor]
}

}
\keyword{internal}
